[v2.1] Step 6: Merge 'Customer Reviews' as additional tab into WebShop
This commit contains the sources where the standalone 'Customer Reviews'
web application is merged as additional tab into the ESPM WebShop
application. This component is part of the ESPM WebShop Application
which is running as demo application on SAP HANA Cloud
https://cloudmodelespmhana.hana.ondemand.com/espm-cloud-web/webshop

The corresponding WebShop Extension Scenario description on SCN
http://scn.sap.com/docs/DOC-41491
is a step by step tutorial for any developers. 
In Chapter 6 it is described how the extended WebShop application is
deployed and run on SAP HANA Cloud

Used technologies SAPUI5, OData, SAP HANA Cloud Connectivity